+++ all/environments/__init__.py
@@ -1,6 +1,5 @@
 from .abstract import Environment
 from .gym import GymEnvironment
 from .atari import AtariEnvironment
-from .state import State
 
-__all__ = ["Environment", "State", "GymEnvironment", "AtariEnvironment"]
+__all__ = ["Environment", "GymEnvironment", "AtariEnvironment"]
