+++ all/policies/soft_deterministic_test.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch_testing as tt
 from gym.spaces import Box
 from all import nn
-from all.environments import State
+from all.core import State
 from all.policies import SoftDeterministicPolicy
 
 STATE_DIM = 2
