+++ all/policies/softmax_test.py
@@ -2,7 +2,7 @@ import unittest
 import torch
 from torch import nn
 import torch_testing as tt
-from all.environments import State
+from all.core import State
 from all.policies import SoftmaxPolicy
 
 STATE_DIM = 2
